requirements.txt
pyaudio
SpeechRecognition
google-genai
python-dotenv